pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-firebolt"
version = "0.1.0"
description = "LangChain integration for Firebolt vector store"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Firebolt Analytics", email = "support@firebolt.io"}
]
keywords = ["langchain", "firebolt", "vectorstore", "embeddings", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain-core>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "firebolt-sdk>=1.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "langchain-tests>=0.0.1",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/firebolt-db/langchain-firebolt"
Documentation = "https://github.com/firebolt-db/langchain-firebolt#readme"
Repository = "https://github.com/firebolt-db/langchain-firebolt"
Issues = "https://github.com/firebolt-db/langchain-firebolt/issues"

[tool.setuptools]
packages = ["langchain_firebolt"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

